requirements.txt
selenium
beautifulsoup4
requests
mysql-connector-python